backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.27.2
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.0.1
aiosqlite==0.20.0
spacy==3.8.11
scikit-learn==1.6.1
joblib==1.5.2
numpy==2.3.2
pandas==2.3.2
pytest==8.3.4
pytest-asyncio==0.24.0
sentence-transformers==3.3.1
